Exercise 5.17 - 5.22

+++ part_5/src/components/Blog.js
@@ -4,21 +4,27 @@ const Blog = ({ blog, handClickLikes, removeBlog }) => {
   const [show, setShow] = useState(false)
   const showWhenVisible = { display: !show ? 'none' : '' }
   const handleDelete = async (id, blog) => {
-    const confirm = window.confirm(`remove blog ${blog.title}`)
-    if (confirm) {
-      try {
-        await blogService.remove(id)
-        removeBlog(blog)
-      }
-      catch (err) {
-        console.log(err)
+    const checkBlog = await blogService.getBlog(id)
+    if(JSON.parse(localStorage.getItem('user')).name.toString() === checkBlog.user.name)
+    {
+      const confirm = window.confirm(`remove blog ${blog.title}`)
+      if (confirm) {
+        try {
+          await blogService.remove(id)
+          removeBlog(blog)
+        }
+        catch (err) {
+          console.log(err)
+        }
       }
     }
   }
   return (
-    <div style={blogStyle}>
-      {blog.title} {blog.author}
-      <button onClick={() => setShow(!show)} style={{ marginLeft: '20px' }}>{show ? 'hide' : 'show'}</button>
+    <div style={blogStyle} className='blogPreview'>
+      <div >
+        {blog.title} {blog.author}
+      </div>
+      <button className='toggle-btn' id='toggle-btn' onClick={() => setShow(!show)} style={{ marginLeft: '20px' }}>{show ? 'hide' : 'show'}</button>
       <div style={showWhenVisible} className='blogContent'>
         <div>{blog.url}</div>
         <div>likes {blog.likes}
@@ -27,7 +33,7 @@ const Blog = ({ blog, handClickLikes, removeBlog }) => {
         </div>
         <div>{blog.user.name}</div>
         <div>
-          <button onClick={() => handleDelete(blog.id, blog)} style={{ backgroundColor: 'blue', cursor: 'pointer' }}>remove</button>
+          <button className='remove-btn' onClick={() => handleDelete(blog.id, blog)} style={{ backgroundColor: 'blue', cursor: 'pointer' }}>remove</button>
         </div>
       </div>
     </div>
